Refactor into focused modules and fix bugs
PhotoMatching.py split into:
- game_types.py: GameType enum, get_game_type, path/CSV constants
- image_utils.py: ensure_valid_image, foil_score, is_probably_foil
- matching.py: _l2_normalize, _cosine_score, find_best_matches
- features.py: MobileNetV2 feature extraction + activation heatmap
- card_database.py: cache build/load, thread-safe _process_image
- csv_manager.py: update_cardlist, get_available_sets, CSV helpers

UI.py split into:
- main_window.py: MainWindow + setup_logging
- settings_window.py: SettingsWindow dialog
- UI.py: entry point only (QApplication + app.exec())

PhotoMatching.py is now a backward-compat re-export shim + CLI.

Bug fixes:
- csv_manager: 5-col rows now use col[3] as count (col[4] tag was wrongly used)
- csv_manager: get_available_sets no longer filters Lorcana sets by isdigit()
- card_database: _process_image receives db_path explicitly (thread safety)
- main_window: removed duplicate worker code and dead code after return
- settings_window: no longer calls set_database_path while listing sets
- .gitignore: fixed DBCardCache_RiftBound.json capitalisation
- UI.py: added missing app.exec() call

Tests: updated 5-col test to match corrected behaviour (47/47 pass)
