+++ .gitignore
@@ -31,6 +31,5 @@ captured_cards/
 # If you keep large image sets out of git (optional)
 # Card_Images/** 
 # !Card_Images/.gitkeep
-DBCardCache_RiftBound.json
+DBCardCache_Riftbound.json
 DBCardCache_Lorcana.json
-__pycache__/PhotoMatching.cpython-313.pyc
