+++ settings_window.py
@@ -1,0 +1,171 @@
+# settings_window.py — Settings dialog for camera, UI, and game/set selection.
+
+import logging
+import os
+from typing import List
+
+from PySide6.QtCore import Qt
+from PySide6.QtWidgets import (
+    QCheckBox,
+    QComboBox,
+    QDialog,
+    QFormLayout,
+    QLabel,
+    QLineEdit,
+    QPushButton,
+    QTreeWidget,
+    QTreeWidgetItem,
+)
+
+from card_database import load_cache, set_database_path
+from csv_manager import get_available_sets
+from game_types import BASE_DATABASE_PATH, GameType
+
+
+class SettingsWindow(QDialog):
+    """
+    Settings dialog for camera and UI options.
+
+    Handles camera selection, foil/rotation/crop preferences, confidence
+    threshold, debug mode, and per-game/set filtering.  Configuration is
+    saved to ui_settings.json via the parent MainWindow on Apply.
+    """
+
+    logger = logging.getLogger("SettingsWindow")
+
+    def __init__(self, parent=None):
+        super().__init__(parent)
+        self.setWindowTitle("Settings")
+        self.setFixedSize(360, 520)
+        self.logger.info("Initializing settings window")
+
+        layout = QFormLayout()
+
+        # Camera index selector (0–4)
+        self.camera_combo = QComboBox()
+        for i in range(5):
+            self.camera_combo.addItem(f"Camera {i}", userData=i)
+        self.camera_combo.setCurrentIndex(min(getattr(parent, "camera_index", 0), 4))
+        layout.addRow("Camera:", self.camera_combo)
+
+        help_label = QLabel("Note: If camera doesn't work, try a different index\nand restart the application.")
+        help_label.setStyleSheet("color: #666; font-size: 10px;")
+        layout.addRow(help_label)
+
+        self.keep_foil_checked = QCheckBox("Keep Foil Checked")
+        self.keep_foil_checked.setChecked(getattr(parent, "keep_foil_checked", False))
+        layout.addRow(self.keep_foil_checked)
+
+        self.rotate_checkbox = QCheckBox("Rotate preview & capture 180°")
+        self.rotate_checkbox.setChecked(getattr(parent, "rotate_display", False))
+        layout.addRow(self.rotate_checkbox)
+
+        self.crop_checkbox = QCheckBox("Scan only inside focus box")
+        self.crop_checkbox.setChecked(getattr(parent, "crop_to_focus", True))
+        layout.addRow(self.crop_checkbox)
+
+        self.confidence_input = QLineEdit(
+            str(int(100 * getattr(parent, "confidence_threshold", 0.90)))
+        )
+        layout.addRow("Confidence Threshold (%):", self.confidence_input)
+
+        self.debug_mode_checkbox = QCheckBox("Enable Debug Mode (heatmap overlay)")
+        self.debug_mode_checkbox.setChecked(getattr(parent, "debug_mode", False))
+        layout.addRow(self.debug_mode_checkbox)
+
+        # Game / set tree
+        self.set_tree = QTreeWidget()
+        self.set_tree.setHeaderHidden(True)
+        self.set_tree.setMinimumHeight(200)
+
+        selected_games = getattr(parent, "selected_games", {})
+        selected_sets = getattr(parent, "selected_sets", {})
+
+        game_folders: List[str] = []
+        if os.path.exists(BASE_DATABASE_PATH):
+            game_folders = [
+                item for item in os.listdir(BASE_DATABASE_PATH)
+                if os.path.isdir(os.path.join(BASE_DATABASE_PATH, item))
+                and item not in ("__pycache__",)
+            ]
+
+        for game_name in game_folders:
+            game_root = QTreeWidgetItem(self.set_tree, [game_name])
+            game_root.setFlags(
+                game_root.flags() | Qt.ItemIsAutoTristate | Qt.ItemIsUserCheckable
+            )
+            is_selected = selected_games.get(game_name.lower(), False)
+            game_root.setCheckState(0, Qt.Checked if is_selected else Qt.Unchecked)
+
+            # Pass db_path explicitly — avoids touching the global databasePath
+            game_db_path = os.path.join(BASE_DATABASE_PATH, game_name)
+            try:
+                game_sets = get_available_sets(db_path=game_db_path)
+                game_selected_sets = selected_sets.get(game_name.lower(), [])
+                for set_code in game_sets:
+                    set_item = QTreeWidgetItem(game_root, [f"Set {set_code}"])
+                    set_item.setFlags(set_item.flags() | Qt.ItemIsUserCheckable)
+                    set_item.setCheckState(
+                        0, Qt.Checked if set_code in game_selected_sets else Qt.Unchecked
+                    )
+                    set_item.setData(0, Qt.UserRole, (game_name.lower(), set_code))
+            except Exception as e:
+                self.logger.warning(f"Error getting sets for {game_name}: {e}")
+
+        self.set_tree.expandAll()
+        layout.addRow("Filter Sets:", self.set_tree)
+
+        apply_button = QPushButton("Apply")
+        apply_button.clicked.connect(self.apply_settings)
+        layout.addRow(apply_button)
+        self.setLayout(layout)
+
+    def apply_settings(self):
+        """Apply settings to parent MainWindow and persist them."""
+        parent = self.parent()
+        if not parent:
+            self.close()
+            return
+
+        parent.keep_foil_checked = self.keep_foil_checked.isChecked()
+        parent.rotate_display = self.rotate_checkbox.isChecked()
+        parent.crop_to_focus = self.crop_checkbox.isChecked()
+        parent.debug_mode = self.debug_mode_checkbox.isChecked()
+        parent.camera_index = self.camera_combo.currentData()
+
+        try:
+            pct = float(self.confidence_input.text())
+            parent.confidence_threshold = max(0.0, min(1.0, pct / 100.0))
+        except Exception:
+            parent.confidence_threshold = 0.90
+
+        # Collect game / set selections from tree
+        selected_games = {"lorcana": False, "riftbound": False}
+        selected_sets = {"lorcana": [], "riftbound": []}
+
+        for i in range(self.set_tree.topLevelItemCount()):
+            root = self.set_tree.topLevelItem(i)
+            game_name = root.text(0).lower()
+            if game_name not in selected_games:
+                continue
+            selected_games[game_name] = root.checkState(0) != Qt.Unchecked
+            if selected_games[game_name]:
+                for j in range(root.childCount()):
+                    child = root.child(j)
+                    if child.checkState(0) == Qt.Checked:
+                        set_code = child.text(0).split(" ")[-1]
+                        selected_sets[game_name].append(set_code)
+
+        self.logger.info(f"Applying settings — games: {selected_games}, sets: {selected_sets}")
+        parent.selected_games = selected_games
+        parent.selected_sets = selected_sets
+
+        # Load the first selected game's database
+        for game_name, is_selected in selected_games.items():
+            if is_selected:
+                set_database_path(game_name.capitalize())
+                parent.featureDB = load_cache()
+                break
+
+        parent.save_settings()
+        self.close()
